+++ evosax/strategies/ars.py
@@ -6,7 +6,7 @@ from ..strategy import Strategy
 from ..utils import GradientOptimizer
 
 
-class Augmented_RS(Strategy):
+class ARS(Strategy):
     def __init__(
         self,
         num_dims: int,
@@ -25,13 +25,13 @@ class Augmented_RS(Strategy):
         self.elite_popsize = int(self.popsize / 2 * self.elite_ratio)
         assert opt_name in ["sgd", "adam", "rmsprop", "clipup"]
         self.optimizer = GradientOptimizer[opt_name](self.num_dims)
-        self.strategy_name = "Augmented_RS"
+        self.strategy_name = "ARS"
 
     @property
     def params_strategy(self) -> chex.ArrayTree:
         """Return default parameters of evolution strategy."""
         es_params = {
-            "sigma_init": 0.1,
+            "sigma_init": 0.03,
             "sigma_decay": 0.999,
             "sigma_limit": 0.01,
             "init_min": 0.0,