+++ evosax/strategies/cma_es.py
@@ -83,7 +83,7 @@ class CMA_ES(Strategy):
             "c_c": c_c,
             "chi_n": chi_n,
             "weights": weights,
-            "sigma_init": 1.0,
+            "sigma_init": 0.065,
             "weights_truncated": weights_truncated,
             "init_min": 0.0,
             "init_max": 0.0,