+++ evosax/strategies/de.py
@@ -5,13 +5,13 @@ from typing import Tuple
 from ..strategy import Strategy
 
 
-class Differential_ES(Strategy):
+class DE(Strategy):
     def __init__(self, num_dims: int, popsize: int):
         """Differential Evolution (Storn & Price, 1997)
         Reference: https://tinyurl.com/4pje5a74"""
         assert popsize > 6
         super().__init__(num_dims, popsize)
-        self.strategy_name = "Differential_ES"
+        self.strategy_name = "DE"
 
     @property
     def params_strategy(self) -> chex.ArrayTree: